project-service/requirements.txt
fastapi>=0.80
uvicorn[standard]>=0.20
sqlalchemy>=1.4
asyncpg>=0.25
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
httpx>=0.23
python-dotenv>=1.0
